+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # Changelog
 
 All notable changes to `laravel-jira-package` will be documented in this file.
+
+## 1.0.0 - 2023-03-28
+
+### Added
+
+- Feature to get the current board.
+- Feature to get the current board's issues.
+- Feature to get the issue's details